Create ImageViewWithMask in common gui
Change base classes of mage view in modules.
