+++ package/PartSeg/segmentation_mask/image_view.py
@@ -3,7 +3,7 @@ from qtpy.QtGui import QPainter, QPen, QColor
 from qtpy.QtWidgets import QToolTip
 
 from PartSeg.common_gui.channel_control import ChannelProperty
-from ..common_gui.stack_image_view import ImageView, ImageCanvas
+from ..common_gui.stack_image_view import ImageViewWithMask, ImageCanvas
 import numpy as np
 
 
@@ -27,7 +27,7 @@ class StackImageCanvas(ImageCanvas):
             painter.drawRect(self.mark_component)
 
 
-class StackImageView(ImageView):
+class StackImageView(ImageViewWithMask):
     image_canvas = StackImageCanvas
 
     def __init__(self, settings, channel_property: ChannelProperty, name: str):
